Added Langevin thermostat; Also changed variable fCallsMin in ProcessSearchJob.h from int to long to avoid warning in compilation
git-svn-id: http://theory.cm.utexas.edu/svn/eon@843 9d962951-508e-48a8-94b5-974f9cfda47c

+++ client/ProcessSearchJob.h
@@ -41,7 +41,7 @@ class ProcessSearchJob : public Job {
         double prefactorsValues[2];
 
         int fCallsSaddle;
-        int fCallsMin;
+        long fCallsMin;
         int fCallsPrefactors;
 };
